# docs/swagger.yaml
basePath: /api/
definitions:
  httpchi.ErrResp:
    properties:
      error:
        type: string
      param:
        type: string
      value:
        type: string
    type: object
  httpchi.MsgResp:
    properties:
      message:
        type: string
    type: object
  task.Request:
    properties:
      description:
        type: string
      due_date:
        type: string
      status:
        type: boolean
      title:
        type: string
    required:
    - description
    - due_date
    - status
    - title
    type: object
  task.Task:
    properties:
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      status:
        type: boolean
      title:
        type: string
    required:
    - description
    - due_date
    - id
    - status
    - title
    type: object
host: localhost:9090
info:
  contact: {}
  description: API for task manager
  title: task-manager API
  version: "1.0"
paths:
  /task:
    post:
      consumes:
      - application/json
      description: 'Creates a task with specified fields: title, description, due
        date, and completion status'
      parameters:
      - description: Data of the new task
        in: body
        name: taskRequest
        required: true
        schema:
          $ref: '#/definitions/task.Request'
      produces:
      - application/json
      responses:
        "201":
          description: Task successfully created
          schema:
            $ref: '#/definitions/task.Task'
        "400":
          description: Incorrect JSON or invalid date format
          schema:
            $ref: '#/definitions/httpchi.ErrResp'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/httpchi.ErrResp'
      summary: creates a new task
      tags:
      - Tasks
  /task/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a task by the specified identifier
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task successfully deleted
          schema:
            $ref: '#/definitions/httpchi.MsgResp'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/httpchi.MsgResp'
      summary: Deletes a task by ID
      tags:
      - Tasks
    get:
      consumes:
      - application/json
      description: Retrieves a task based on the provided identifier
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Task successfully retrieved
          schema:
            $ref: '#/definitions/task.Task'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/httpchi.MsgResp'
      summary: Gets a task by ID
      tags:
      - Tasks
    put:
      consumes:
      - application/json
      description: Updates a task by the specified identifier
      parameters:
      - description: Task ID
        in: path
        name: id
        required: true
        type: string
      - description: Data for updating the task
        in: body
        name: taskUpd
        required: true
        schema:
          $ref: '#/definitions/task.Request'
      produces:
      - application/json
      responses:
        "200":
          description: Task successfully updated
          schema:
            $ref: '#/definitions/task.Task'
        "400":
          description: Incorrect JSON or invalid date format
          schema:
            $ref: '#/definitions/httpchi.ErrResp'
        "404":
          description: Task not found
          schema:
            $ref: '#/definitions/httpchi.MsgResp'
        "422":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/httpchi.ErrResp'
      summary: Updates a task by ID
      tags:
      - Tasks
  /tasks:
    get:
      consumes:
      - application/json
      description: Retrieves a list of tasks based on status, date, and page for pagination
      parameters:
      - description: Task completion status (true/false)
        in: query
        name: status
        type: string
      - description: 'Task date (format: YYYY-MM-DD)'
        in: query
        name: date
        type: string
      - description: Page number for pagination
        in: query
        name: page
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of tasks
          schema:
            items:
              $ref: '#/definitions/task.Task'
            type: array
        "400":
          description: Invalid request parameters
          schema:
            $ref: '#/definitions/httpchi.ErrResp'
        "404":
          description: Tasks not found
          schema:
            $ref: '#/definitions/httpchi.MsgResp'
      summary: Returns a list of tasks with filtering and pagination
      tags:
      - Tasks
securityDefinitions:
  BasicAuth:
    type: basic
swagger: "2.0"